+++ control-planes/mgmt_api/src/api/v1/mod.rs
@@ -6,7 +6,7 @@ macro_rules! v1_crud_api {
 
     use std::time::Duration;
     use actix_web::{web::{self}, HttpResponse, put, get, delete};
-    use super::models::{ResourceDto, ResourceProviderDto, ReadyWaitParams};
+    use super::models::{ResourceDto, ReadyWaitParams};
 
     pub fn configure(cfg: &mut web::ServiceConfig) {
       cfg
@@ -95,9 +95,8 @@ macro_rules! v1_crud_api {
 macro_rules! v1_crud_api_provider {
   ($domain_service:ty, $spec_dto:ty $(, $additional_route:expr ),* ) => {
 
-    use std::time::Duration;
     use actix_web::{web::{self}, HttpResponse, put, get, delete};
-    use super::models::{ResourceDto, ResourceProviderDto, ReadyWaitParams};
+    use super::models::ResourceProviderDto;
 
     pub fn configure(cfg: &mut web::ServiceConfig) {
       cfg
@@ -111,78 +110,6 @@ macro_rules! v1_crud_api_provider {
         ;
 
     }
-
-
-
-    use serde_json::Value;
-    #[put("/{id}")]
-    async fn put_resource(svc: web::Data<$domain_service>, id: web::Path<String>, body: web::Json<Value>) -> HttpResponse {
-      log::debug!("put_provider_resource: {:?}", id);
-      let data = body.into_inner();
-
-      match svc.set(&id.into_inner(), data).await {
-        Ok(res) => HttpResponse::Ok().json(res),
-        Err(e) => e.into()
-      }
-    }
-
-    #[delete("/{id}")]
-    async fn delete_resource(svc: web::Data<$domain_service>, id: web::Path<String>) -> HttpResponse {
-      log::debug!("deregister_provider_resource: {:?}", id);
-
-      match svc.delete(&id.into_inner()).await {
-        Ok(_) => HttpResponse::NoContent().finish(),
-        Err(e) => e.into()
-      }
-    }
-
-    #[get("")]
-    async fn list_resources(svc: web::Data<$domain_service>) -> HttpResponse {
-      log::debug!("list_resources");
-
-      match svc.list().await {
-        Ok(res) => {
-          let result = res.into_iter().map(|r| ResourceProviderDto::<$spec_dto>::from(r)).collect::<Vec<_>>();
-          HttpResponse::Ok().json(result)
-        },
-        Err(e) => e.into()
-      }
-    }
-
-    #[get("/{id}")]
-    async fn get_resource(svc: web::Data<$domain_service>, id: web::Path<String>) -> HttpResponse {
-      log::debug!("get_resource: {:?}", id);
-
-      match svc.get(&id.into_inner()).await {
-        Ok(res) => HttpResponse::Ok().json(res),
-        Err(e) => e.into()
-      }
-    }
-
-  }
-}
-
-macro_rules! v1_crud_api_provider {
-  ($domain_service:ty, $spec_dto:ty $(, $additional_route:expr ),* ) => {
-
-    use std::time::Duration;
-    use actix_web::{web::{self}, HttpResponse, put, get, delete};
-    use super::models::{ResourceDto, ResourceProviderDto, ReadyWaitParams};
-
-    pub fn configure(cfg: &mut web::ServiceConfig) {
-      cfg
-        .service(put_resource)
-        .service(get_resource)
-        .service(delete_resource)
-        .service(list_resources)
-        $(
-          .service($additional_route)
-        ),*
-        ;
-
-    }
-
-
 
     use serde_json::Value;
     #[put("/{id}")]
@@ -263,7 +190,7 @@ pub mod reaction_handlers {
 pub mod source_provider_handlers {
     use crate::{
         api::v1::models::SourceProviderSpecDto,
-        domain::resource_services::SourceProviderDomainService,
+        domain::resource_provider_services::SourceProviderDomainService,
     };
     v1_crud_api_provider!(SourceProviderDomainService, SourceProviderSpecDto);
 }
@@ -271,7 +198,7 @@ pub mod source_provider_handlers {
 pub mod reaction_provider_handlers {
     use crate::{
         api::v1::models::ReactionProviderSpecDto,
-        domain::resource_services::ReactionProviderDomainService,
+        domain::resource_provider_services::ReactionProviderDomainService,
     };
     v1_crud_api_provider!(ReactionProviderDomainService, ReactionProviderSpecDto);
 }
