+++ control-planes/mgmt_api/src/domain/mod.rs
@@ -1,5 +1,6 @@
 pub mod debug_service;
 pub mod mappings;
 pub mod models;
+pub mod resource_provider_services;
 pub mod resource_services;
 pub mod result_service;
